feat: Show tip on first request and refactor phrases
- Shows an informative tip on the first request to help new users.
- Refactors witty phrases and tips into separate constant files for better organization.
- Updates tests to reflect the new behavior and improve robustness.
